requirements.txt
streamlit
pydantic
google-genai
python-dotenv
PyMuPDF
python-docx
psycopg2-binary